Use star icon to toggle favourite in Favourites activity
Since it is possible to add favourites in Kore, it should also be possible
to remove them. Instead of introducing the context menu in the Favourites
activity, this change introduces a star button next to the context menu.

The star is intended to replace the context menu entry "Add to favourites"
as well, in a later change.

The star icon has built-in user feedback by toggeling between a full and
outline image, so feedback via toasts is unnecessary.